% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_line_neighbors.R
\name{blank_line_neighbors}
\alias{blank_line_neighbors}
\title{Identify blank lines around a line}
\usage{
blank_line_neighbors(
  x,
  pattern = NULL,
  index = NULL,
  top = TRUE,
  bottom = TRUE
)
}
\arguments{
\item{x}{A \code{\link[base:character]{character}} vector.}

\item{pattern}{A string indicating the pattern to look for. If there's more
than one match, the first one will be used (default: \code{NULL}).}

\item{index}{An integer number indicating the line to look for (default:
\code{NULL}). If this parameter is provided, the \code{pattern} parameter will be
ignored.}

\item{top}{A \code{\link[base:logical]{logical}} flag indicating if the function
should look for blank lines at the top of the line (default: \code{TRUE}).}

\item{bottom}{A \code{\link[base:logical]{logical}} flag indicating if the
function should look for blank lines at the bottom of the line (default:
\code{TRUE}).}
}
\value{
An \code{\link[=as.integer]{integer}} vector with the indexes of the blank
lines.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{blank_line_neighbors()} identify and returns the indexes of blank lines
around a specific line matching a pattern.
}
\examples{
file <- tempfile()
writeLines(c("", "line 1", "", "", "line 2", "", "", "line 3", ""), file)
x <- readLines(file)

blank_line_neighbors(x, pattern = "^line 2$")
#> [1] 4 6 # Expected
}
\seealso{
Other file functions: 
\code{\link{normalize_hashtags}()},
\code{\link{normalize_names}()},
\code{\link{peek_csv_file}()},
\code{\link{remove_blank_line_dups}()},
\code{\link{replace_in_file}()},
\code{\link{split_file}()}
}
\concept{file functions}
