% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_vector.R
\name{cut_vector}
\alias{cut_vector}
\title{Cut a vector into segments}
\usage{
cut_vector(
  x,
  index = NULL,
  n = NULL,
  between = NULL,
  rm_start = FALSE,
  rm_end = FALSE,
  round_fun = floor
)
}
\arguments{
\item{x}{An \link[checkmate:checkAtomicVector]{atomic} vector.}

\item{index}{(optional) An \link[checkmate:checkIntegerish]{integerish}
\code{numeric} object with the indexes/cutting points (default: \code{NULL}).}

\item{n}{(optional) An \link[checkmate:checkIntegerish]{integerish} number
specifying the number of segments to cut the vector into. If \code{n} is
specified, the \code{index} argument is ignored (default: \code{NULL}).}

\item{between}{(optional) A string specifying the direction of the cut:
\code{"left"} or \code{"right"}. Use this argument only when the cut should be
performed between the index values (default: \code{NULL}).}

\item{rm_start}{(optional) a \code{\link[base:logical]{logical}} flag indicating if
the start element of the cut must be removed (default: \code{FALSE}).}

\item{rm_end}{(optional) a \code{\link[base:logical]{logical}} flag indicating if
the end element of the cut must be removed (default: \code{FALSE}).}

\item{round_fun}{(optional) A function to round the number of elements in
when cutting using \code{n} (default: \code{floor}).}
}
\value{
A \code{\link[base:list]{list}} object with the cut pieces as elements.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{cut_vector()} splits a vector into segments at specified cutting points
or indexes. You can define the cuts either by providing explicit index
positions or by specifying the number of segments.

Additional options allow you to control whether cuts occur between or at
index values, and whether to remove elements at the start or end of the
resulting segments.
}
\details{
\code{cut_vector()} can perform different kinds of cuts. Here are some examples.
\subsection{Cutting by index values}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{cut_vector(seq(10), c(3, 9))

   cut         cut
    |           |
1 2 3 4 5 6 7 8 9 10

Element 1: 1, 2
Element 2: 4, 5, 6, 7, 8
Element 3: 10
}\if{html}{\out{</div>}}
}

\subsection{Cutting between index values}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{cut_vector(seq(10), c(3, 9), between = "left")

  cut         cut
   |           |
1 2 3 4 5 6 7 8 9 10

Element 1: 1, 2
Element 2: 3, 4, 5, 6, 7, 8
Element 3: 9, 10
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{cut_vector(seq(10), c(3, 9), between = "right")

    cut         cut
     |           |
1 2 3 4 5 6 7 8 9 10

Element 1: 1, 2, 3
Element 2: 4, 5, 6, 7, 8, 9
Element 3: 10
}\if{html}{\out{</div>}}
}

\subsection{Cutting by number of segments}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{cut_vector(seq(10), n = 3, round_fun = floor)

    cut   cut
     |     |
1 2 3 4 5 6 7 8 9 10

Element 1: 1, 2, 3
Element 2: 4, 5, 6
Element 3: 7, 8, 9, 10
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{cut_vector(seq(10), n = 3, round_fun = ceiling)

      cut     cut
       |       |
1 2 3 4 5 6 7 8 9 10

Element 1: 1, 2, 3, 4
Element 2: 5, 6, 7, 8
Element 3: 9, 10
}\if{html}{\out{</div>}}
}

\subsection{Removing tips}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{cut_vector(seq(20), c(7, 16), rm_start = TRUE, rm_end = TRUE)

           cut                      cut
            |                        |
1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20
|---------|                            |---------|
 start tip                               end tip

Element 1: 8, 9, 10, 11, 12, 13, 14, 15
}\if{html}{\out{</div>}}
}
}
\examples{
## Cutting by index values

cut_vector(seq(10), c(3, 9))

## Cutting between index values

cut_vector(seq(10), c(3, 9), between = "left")

cut_vector(seq(10), c(3, 9), between = "right")

## Removing start or end tips

cut_vector(seq(10), c(3, 9), rm_start = TRUE)

cut_vector(seq(10), c(3, 9), rm_end = TRUE)
}
\seealso{
Other vector functions: 
\code{\link{remove_caps}()},
\code{\link{split_by_pattern}()}
}
\concept{vector functions}
