% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_match_dbl_2.R
\name{find_closest_match_dbl_2}
\alias{find_closest_match_dbl_2}
\title{Find closest matches between numeric vectors}
\usage{
find_closest_match_dbl_2(x, y, x_ref, y_ref, tol = 0.05)
}
\arguments{
\item{x}{A \code{\link[base:numeric]{numeric}} vector.}

\item{y}{A \code{\link[base:numeric]{numeric}} vector.}

\item{x_ref}{A \code{\link[base:numeric]{numeric}} vector with reference values for
\code{x}.}

\item{y_ref}{A \code{\link[base:numeric]{numeric}} vector with reference values for
\code{y}.}

\item{tol}{(Optional) A \code{\link[base:numeric]{numeric}} value that sets the
maximum acceptable difference for a match (Default: \code{0.05}).}
}
\value{
If any value was found, returns the index of the closest match in
\code{x_ref}/\code{y_ref}. Else, returns \code{NA}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{find_closest_match_dbl_2()} identifies the closest matches between two
\code{\link[base:numeric]{numeric}} vectors.

Matches are determined by calculating the absolute differences between
reference and target values. A user-defined tolerance sets the maximum
acceptable difference for a match. The closest match is the pair with
the smallest absolute difference that falls within the specified tolerance.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(1.1, 2.1, 3.1, 4.1, 5.1)
x_ref <- c(5, 1, 3, 4, 2)
y_ref <- c(5, 1, 3, 4, 2)
tol <- 0.5

find_closest_match_dbl_2(x, y, x_ref, y_ref, tol)
#> [1] 2 5 3 4 1 # Expected
}
\seealso{
Other match functions: 
\code{\link{match_strings}()}
}
\concept{match functions}
