% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_hms.R
\name{fmt_hms}
\alias{fmt_hms}
\title{Format time duration strings into \code{hms} for \code{gtsummary} objects}
\usage{
fmt_hms(x, big_mark = ",", decimal_mark = ".")
}
\arguments{
\item{x}{A \code{\link[gtsummary:as_gtsummary]{gtsummary}} object.}

\item{big_mark}{(optional) A \code{\link[base:character]{character}} string indicating
the thousands separator (default: \code{","}).}

\item{decimal_mark}{(optional) A \code{\link[base:character]{character}} string
indicating the decimal separator (default: \code{"."}).}
}
\value{
A \code{\link[gtsummary:as_gtsummary]{gtsummary}} object with formatted
\code{hms} strings in all \code{stat_} columns.
}
\description{
\code{fmt_hms()} formats strings representing time durations
(e.g.,\code{"1,234.56 secs"}) in \code{\link[gtsummary:as_gtsummary]{gtsummary}} objects
to a more readable \code{\link[hms:hms]{hms}} format (e.g., \code{"00:20:34"}).
}
\examples{
library(dplyr)
library(gtsummary)
library(hms)

table <-
  tibble(
    time =
      seq(
        from = parse_hms("00:00:00"),
        to = parse_hms("23:59:59"),
        by = parse_hms("00:00:01")
      ) |>
      sample(1000)
  ) |>
  tbl_summary(
    label = list(time ~ "Time of Day"),
    type = all_continuous() ~ "continuous2",
    statistic = all_continuous2() ~ c("{mean} ({sd})", "{min}, {max}")
  )

table

table |> fmt_hms()
}
\concept{\code{gtsummary} functions}
