% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shared_dir.R
\name{get_shared_dir}
\alias{get_shared_dir}
\title{Get the Shared Directory of Files}
\usage{
get_shared_dir(files, expand_paths = FALSE)
}
\arguments{
\item{files}{A \code{\link[base:character]{character}} vector of file paths.}

\item{expand_paths}{(optional) A \code{\link[base:logical]{logical}} flag
indicating whether to expand relative paths to absolute paths. If \code{TRUE},
all paths must exist (default: \code{FALSE}).}
}
\value{
A \code{\link[base:character]{character}} vector of shared directories.
If no shared directories are found, it returns \code{NA_character_}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{get_shared_dir()} identifies the common parent directories among a set of
file paths. It returns the directories that are shared by two or more files
in the input vector.
}
\examples{
get_shared_dir(c("/home/user/file1.txt", "/home/user/file2.txt"))
#> [1] "/home/user"

files <- c(
  "/home/user/file1.txt",
  "/home/user/file2.txt",
  "/home/user/dir/file3.txt",
  "/home/test.txt",
  "C:\\\\Users\\\\user\\\\file1.txt"
)

get_shared_dir(files)
#> [1] "/home"      "/home/user"
}
\seealso{
Other file functions: 
\code{\link{identify_blank_line_neighbors}()},
\code{\link{normalize_hashtags}()},
\code{\link{normalize_names}()},
\code{\link{peek_csv_file}()},
\code{\link{remove_blank_line_dups}()},
\code{\link{replace_in_file}()},
\code{\link{sort_files_by_size}()},
\code{\link{split_file}()},
\code{\link{split_files_by_size}()},
\code{\link{zip_files_by_pattern}()}
}
\concept{file functions}
