% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groomr-package.R
\docType{package}
\name{groomr-package}
\alias{groomr}
\alias{groomr-package}
\title{groomr: Tidy Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of tools for tidying and organizing R objects and files. It provides functions for cleaning, renaming, splitting, and managing files and data, making workflows more efficient and reproducible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/groomr/}
  \item \url{https://github.com/danielvartan/groomr/}
  \item Report bugs at \url{https://github.com/danielvartan/groomr/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
