% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_machine_readable.R
\name{make_machine_readable}
\alias{make_machine_readable}
\title{Make a string machine-readable}
\usage{
make_machine_readable(x)
}
\arguments{
\item{x}{A \code{\link[base:character]{character}} vector.}
}
\value{
A \code{\link[base:character]{character}} vector with machine-readable names.
}
\description{
\code{make_machine_readable()} makes a string machine-readable by converting it
to lowercase, removing special characters, and replacing spaces with hyphens.

This function follows the naming conventions from the
\href{https://style.tidyverse.org/files.html#names}{Tidyverse Style Guide}.
}
\examples{
make_machine_readable("Test - Test ")
#> [1] "test-test" # Expected

make_machine_readable("!@#TéSt$$\%^")
#> [1] "test" # Expected

make_machine_readable("TeSt_test.txt")
#> [1] "test-test.txt" # Expected
}
\seealso{
Other string functions: 
\code{\link{to_ascii}()},
\code{\link{to_title_case_pt}()}
}
\concept{string functions}
