% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_extdata.R
\name{normalize_extdata}
\alias{normalize_extdata}
\title{Normalize file names from the \code{extdata} folder}
\usage{
normalize_extdata(package = basename(here::here()))
}
\arguments{
\item{package}{(optional) a string indicating the target package (default:
\code{basename(here::here())}).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{ATTENTION}: This function must be used only with
\href{https://posit.co/products/open-source/rstudio/}{RStudio} IDE.

\code{normalize_extdata()} normalize file names founded in the \verb{./inst/extdata/}
folder of a R package.

The function lowercase the names, substitute \code{" - "} characters for \code{"_"},
and substitute \code{" "} characters for \code{"-"}.
}
\examples{
\dontrun{
normalize_extdata()}
}
\concept{R package functions}
