% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_html_table.R
\name{parse_html_table}
\alias{parse_html_table}
\title{Parse HTML tables}
\usage{
parse_html_table(x = clipr::read_clip())
}
\arguments{
\item{x}{(optional) A \code{\link[base:character]{character}} vector with HTML code.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with the parsed HTML table.
}
\description{
\code{parse_html_table()} parses a HTML table to a \code{\link[tibble:tibble]{tibble}}.
}
\examples{
x <- c(
  "<table>",
  "  <tr>",
  "    <th>Name</th>",
  "    <th>Age</th>",
  "  </tr>",
  "  <tr>",
  "    <td>Alice</td>",
  "    <td>30</td>",
  "  </tr>",
  "  <tr>",
  "    <td>Bob</td>",
  "    <td>25</td>",
  "  </tr>",
  "</table>"
)

x |> parse_html_table()
}
\concept{parse functions}
