% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_blank_line_dups.R
\name{remove_blank_line_dups}
\alias{remove_blank_line_dups}
\title{Remove blank line duplicates from a file}
\usage{
remove_blank_line_dups(
  file,
  remove_caps = TRUE,
  block_fix = TRUE,
  block_start_pattern = "<!-- badges: start -->",
  block_end_pattern = "<!-- badges: end -->"
)
}
\arguments{
\item{file}{A string indicating a file path.}

\item{remove_caps}{A \code{\link[=as.logical]{logical}} flag indicating if the function
should remove blank lines at the top and bottom of the file.}

\item{block_fix}{A \code{\link[=as.logical]{logical}} flag indicating if the function
should fix blank lines around a text block}

\item{block_start_pattern}{A string indicating the start of the text block.}

\item{block_end_pattern}{A string indicating the end of text block.}
}
\value{
An invisible \code{NULL}. This function don't aim to return values.
}
\description{
This function removes blank line duplicates (e.g., a \code{""} line, followed by
another \code{""} line) from a file.
}
\details{
It can also remove blank lines at the top or bottom of the file, and remove
blank lines around a text block.
}
\examples{
file <- tempfile()
writeLines(c("", "line 1", "", "", "line 2", "", "", "line 3", ""), file)

remove_blank_line_dups(file)

readLines(file)
#> [1] "line 1" ""       "line 2" ""       "line 3" # Expected
}
