% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_caps.R
\name{remove_caps}
\alias{remove_caps}
\title{Remove the first and/or last element of a vector}
\usage{
remove_caps(x, start = TRUE, end = TRUE)
}
\arguments{
\item{x}{An \code{\link[base:is.recursive]{atomic}} vector.}

\item{start}{(optional) A \code{\link[base:logical]{logical}} flag indicating whether
to remove the first element of \code{x} (default: \code{TRUE}).}

\item{end}{(optional) A \code{\link[base:logical]{logical}} flag indicating whether
to remove the last element of \code{x} (default: \code{TRUE}).}
}
\value{
The same vector \code{x} with the first and/or last element removed.
}
\description{
\code{remove_caps()} removes the first and/or last element of a vector.
}
\examples{
remove_caps(c("a", "b", "c"))
#> [1] "b" # Expected

remove_caps(c("a", "b", "c"), start = FALSE)
#> [1] "a" "b" # Expected

remove_caps(c("a", "b", "c"), end = FALSE)
#> [1] "b" "c" # Expected
}
\seealso{
Other vector functions: 
\code{\link{cut_vector}()},
\code{\link{split_by_pattern}()}
}
\concept{vector functions}
