% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_caps.R
\name{replace_caps}
\alias{replace_caps}
\title{Replace the first and/or last element of a vector}
\usage{
replace_caps(x, replacement = NA, start = TRUE, end = TRUE)
}
\arguments{
\item{x}{An \code{\link[base:is.recursive]{atomic}} vector.}

\item{replacement}{(optional) A value to replace the first and/or last
element of \code{x} with (default: \code{NA}).}

\item{start}{(optional) A \code{\link[base:logical]{logical}} flag indicating whether
to remove the first element of \code{x} (default: \code{TRUE}).}

\item{end}{(optional) A \code{\link[base:logical]{logical}} flag indicating whether
to remove the last element of \code{x} (default: \code{TRUE}).}
}
\value{
The same vector \code{x} with the first and/or last element replaced.
}
\description{
\code{replace_caps()} replaces the first and/or last element of a vector.
}
\examples{
replace_caps(c("a", "b", "c"))
#> [1] NA "b" NA # Expected

replace_caps(c("a", "b", "c"), start = FALSE)
#> [1] "a" "b" NA # Expected

replace_caps(c("a", "b", "c"), end = FALSE)
#> [1] "b" "c" NA # Expected
}
\seealso{
Other vector functions: 
\code{\link{cut_vector}()},
\code{\link{remove_caps}()},
\code{\link{split_by_pattern}()}
}
\concept{vector functions}
