% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_caps.R
\name{rm_caps}
\alias{rm_caps}
\title{Remove the first and/or last element of a `character`` vector}
\usage{
rm_caps(x, start = TRUE, end = TRUE)
}
\arguments{
\item{x}{A \code{\link[base:character]{character}} vector.}

\item{start}{A \code{\link[base:logical]{logical}} flag indicating whether to remove
the first element of \code{x} (Default: \code{TRUE}).}

\item{end}{A \code{\link[base:logical]{logical}} flag indicating whether to remove
the last element of \code{x} (Default: \code{TRUE}).}
}
\value{
A \code{\link[base:character]{character}} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{rm_caps()} removes the first and/or last element of a `character`` vector.
}
\examples{
rm_caps(c("a", "b", "c"))
#> [1] "b" # Expected

rm_caps(c("a", "b", "c"), start = FALSE)
#> [1] "a" "b" # Expected

rm_caps(c("a", "b", "c"), end = FALSE)
#> [1] "b" "c" # Expected
}
\seealso{
Other vector functions: 
\code{\link{cutter}()},
\code{\link{split_by_pattern}()}
}
\concept{vector functions}
