% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_pattern.R
\name{split_by_pattern}
\alias{split_by_pattern}
\title{Split a character vector by a pattern}
\usage{
split_by_pattern(
  x,
  start_pattern = "^# |^## |---",
  end_pattern = NULL,
  name_list = TRUE,
  include_start = FALSE
)
}
\arguments{
\item{x}{A \code{\link[base:character]{character}} vector}

\item{start_pattern}{A string with the start pattern (default:
\code{"^# |^## |---"})}

\item{end_pattern}{A string with the end pattern. Use \code{NULL} if there isn't
one (default: \code{NULL}).}

\item{name_list}{A \code{\link[base:logical]{logical}} flag indicating if each
output item should have a name (default: \code{TRUE}).}

\item{include_start}{A \code{\link[base:logical]{logical}} flag indicating if the
line flagged in the start pattern should be included in the split (default:
\code{FALSE}).}
}
\value{
A list with the split character vector.
}
\description{
\code{split_by_pattern()} allow you to split a character vector considering a
start and end pattern.
}
\examples{
x <- c(
 "# Title 1",
 "## Subtitle 1",
 "Content 1",
 "Content 2",
 "# Title 2",
 "## Subtitle 2",
 "Content 3",
 "Content 4"
 )

 split_by_pattern(
   x,
   start_pattern = "^# |^## |---",
   end_pattern = NULL,
   name_list = TRUE,
   include_start = FALSE
 )
 #> $`# Title 1 [line 1]` # Expected
 #> character(0) # Expected
 #> # Expected
 #> $`## Subtitle 1 [line 2]` # Expected
 #> [1] "Content 1" "Content 2" # Expected
 #> # Expected
 #> $`# Title 2 [line 5]` # Expected
 #> character(0) "# Title 2" # Expected
 #> # Expected
 #> $`## Subtitle 2 [line 6]` # Expected
 #> [1] "Content 3" "Content 4" # Expected
}
\seealso{
Other string functions: 
\code{\link{replace_pattern}()}
}
\concept{string functions}
