% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groomr-package.R
\docType{package}
\name{groomr-package}
\alias{groomr}
\alias{groomr-package}
\title{groomr: Tools to Tidy Strings and Text Files}
\description{
A set of tools to tidy text files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielvartan/groomr/}
  \item Report bugs at \url{https://github.com/danielvartan/groomr/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
