% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_names.R
\name{normalize_names}
\alias{normalize_names}
\title{Make file and directory names machine-readable}
\usage{
normalize_names(
  path = clipr::read_clip(),
  regex = "^README|^OFL.txt$|^DESCRIPTION|^Google README.txt$",
  invert = TRUE,
  ...
)
}
\arguments{
\item{path}{(Optional) A \code{\link[base:character]{character}} string indicating
the directory path. Defaults to clipboard content.}

\item{regex}{(Optional) A \code{\link[base:character]{character}} string with
a regular expression to filter files/directories to rename
(Default: \code{"^README|^OFL.txt$|^DESCRIPTION|^Google README.txt$"}).}

\item{invert}{(Optional) A \code{\link[base:logical]{logical}} flag indicating if
the regular expression in \code{regex} should be negated (Default: \code{TRUE}).}

\item{...}{(Optional) Additional arguments to be passed to
\code{\link[fs:dir_ls]{dir_ls()}} while listing files/directories to rename.}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{normalize_names()} normalizes file and directory names to make them
machine-readable.

This function uses \code{\link[fs:dir_ls]{dir_ls()}} to list files/directories to
rename, along with \code{\link[=make_machine_readable]{make_machine_readable()}}
to rename files.
}
\examples{
dir <- tempfile("")
dir.create(dir) |> invisible()

file <- tempfile("Test - Test-", tmpdir = dir)
file.create(file) |> invisible()

normalize_names(dir)

list.files(dir)
#> [1] "test-test-***" # Expected
}
\seealso{
Other file functions: 
\code{\link{blank_line_neighbors}()},
\code{\link{normalize_hashtags}()},
\code{\link{peek_csv_file}()},
\code{\link{remove_blank_line_dups}()},
\code{\link{replace_in_file}()},
\code{\link{split_file}()}
}
\concept{file functions}
