% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek_csv_file.R
\name{peek_csv_file}
\alias{peek_csv_file}
\title{Peek the dimensions of a CSV file}
\usage{
peek_csv_file(file, delim = ",", skip = 0, has_header = TRUE)
}
\arguments{
\item{file}{A string specifying the path to the file.}

\item{delim}{A string specifying the delimiter used in the file
(default: \code{","}).}

\item{skip}{An \code{\link[base:integer]{integer}} specifying the number of
lines to skip before reading the file (default: \code{0}).}

\item{has_header}{A \code{\link[base:logical]{logical}} flag indicating whether
the file has a header row. If \code{TRUE}, the first row is
treated as column names (default: \code{TRUE}).}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{peek_csv_file()} reads a CSV file and provides a summary of its contents,
including the number of columns, rows, and cells. It is useful for quickly
inspecting the structure of a CSV file without loading the entire dataset
into memory.
}
\examples{
library(datasets)
library(readr)

file <- tempfile()

write_csv(mtcars, file)

peek_csv_file(file)
#> The file has 11 columns, 32 rows*, and 352 cells. # Expected
#>
#> # Note: Row count excludes any skipped lines.
}
\seealso{
Other file functions: 
\code{\link{blank_line_neighbors}()},
\code{\link{normalize_hashtags}()},
\code{\link{normalize_names}()},
\code{\link{remove_blank_line_dups}()},
\code{\link{replace_in_file}()},
\code{\link{split_file}()}
}
\concept{file functions}
