% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_in_file.R
\name{replace_in_file}
\alias{replace_in_file}
\title{Replace patterns inside files}
\usage{
replace_in_file(dir = clipr::read_clip(), pattern, replacement)
}
\arguments{
\item{dir}{(Optional) a string indicating the directory of the files.
Defaults to clipboard content.}

\item{pattern}{A string indicating the pattern to look for. The default
interpretation is a regular expression. This parameter will be used on
\code{\link[stringr:str_replace]{stringr::str_replace_all()}}.}

\item{replacement}{A string indicating the replacement value. This parameter
will be used on \code{\link[stringr:str_replace]{stringr::str_replace_all()}}.}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{replace_in_file()} searches each line of a file and replaces a specific
pattern when found.
}
\examples{
file_name <- tempfile(tmpdir = tempfile())
dir_name <- dirname(file_name)
dir.create(dir_name)
file.create(file_name)

con <- file(file_name, "r+")
data <- c("01:00:56", "", "Test", "")
writeLines(data, con = con)
close(con)

replace_in_file(
  dir = dir_name,
  pattern = "([0-1][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]",
  replacement = ""
  )

con <- file(file_name, "r+")
readLines(con)
close(con)
}
\seealso{
Other string functions: 
\code{\link{make_machine_readable}()},
\code{\link{match_strings}()},
\code{\link{to_ascii}()},
\code{\link{to_title_case_pt}()}

Other file functions: 
\code{\link{blank_line_neighbors}()},
\code{\link{normalize_hashtags}()},
\code{\link{normalize_names}()},
\code{\link{remove_blank_line_dups}()},
\code{\link{split_file}()}
}
\concept{file functions}
\concept{string functions}
