% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_file.R
\name{split_file}
\alias{split_file}
\title{Split files into parts}
\usage{
split_file(file = file.choose(), n, dir = dirname(file), has_header = TRUE)
}
\arguments{
\item{file}{(optional) a string indicating the file path.
(default:: \code{base::file.choose()}).}

\item{n}{An \link[checkmate:checkIntegerish]{integerish} indicating the
amount of parts to split the file.}

\item{dir}{(optional) a string indicating the directory where to write the
file parts. (default:: \code{base::dirname(file)}).}

\item{has_header}{(optional) a \code{\link[base:logical]{logical}} flag
indicating if the file has a header (e.g., a CSV file). (default: \code{TRUE}).}
}
\value{
An invisible \code{NULL}. This function don't aim to return values.
}
\description{
\code{split_file()} split a single file into \code{n} parts.
}
\examples{
file_name <- tempfile(tmpdir = tempfile())
dir_name <- dirname(file_name)
dir.create(dir_name)
file.create(file_name)

con <- file(file_name, "r+")
data <- as.character(0:100)
writeLines(data, con = con)
close(con)

split_file(file = file_name, n = 3, has_header = FALSE)

con <- file(file_name, "r+")
readLines(con)
close(con)
}
\concept{file functions.}
